updated the code according to https://review.openbloomberg.com/D28#comment-1

+++ groups/bsl/bslstl/bslstl_stringrefdata.t.cpp
@@ -534,7 +534,8 @@ ASSERT(&*strObj.end()   == strRf2.end());
             Y = X;
 
             LOOP_ASSERT(LINE, Y.begin() == STR);
-            LOOP_ASSERT(LINE, Y.end() - Y.begin() == (int) strlen(STR));
+            LOOP_ASSERT(LINE, Y.end() - Y.begin()
+                                       == static_cast<ptrdiff_t>(strlen(STR)));
             LOOP_ASSERT(LINE, *Y.end() == '\0');
         }
 
@@ -720,7 +721,8 @@ ASSERT(&*strObj.end()   == strRf2.end());
             bslstl::StringRefData<char> Y(X);
 
             LOOP_ASSERT(LINE, Y.begin() == STR);
-            LOOP_ASSERT(LINE, Y.end() - Y.begin() == (int) strlen(STR));
+            LOOP_ASSERT(LINE, Y.end() - Y.begin()
+                                       == static_cast<ptrdiff_t>(strlen(STR)));
             LOOP_ASSERT(LINE, *Y.end() == '\0');
         }
       } break;
@@ -963,7 +965,8 @@ ASSERT(&*strObj.end()   == strRf2.end());
 
             bslstl::StringRefData<char> Y(STR, STR + strlen(STR));
             LOOP_ASSERT(LINE, Y.begin() == STR);
-            LOOP_ASSERT(LINE, Y.end() - Y.begin() == (int) strlen(STR));
+            LOOP_ASSERT(LINE, Y.end() - Y.begin()
+                                       == static_cast<ptrdiff_t>(strlen(STR)));
             LOOP_ASSERT(LINE, *Y.end() == '\0');
         }
 
@@ -1120,7 +1123,8 @@ ASSERT(&*strObj.end()   == strRf2.end());
 
             bslstl::StringRefData<char> Y(STR, STR + strlen(STR));
             LOOP_ASSERT(LINE, Y.begin() == STR);
-            LOOP_ASSERT(LINE, (size_t)(Y.end() - Y.begin()) == strlen(STR));
+            LOOP_ASSERT(LINE, Y.end() - Y.begin()
+                                       == static_cast<ptrdiff_t>(strlen(STR)));
             LOOP_ASSERT(LINE, *Y.end() == '\0');
         }
